+++ test/src/detail_tests/unit_t_operator_tests.cc
@@ -139,21 +139,18 @@ TEST_CASE(
   STATIC_REQUIRE(result == 2);
 }
 
-/// @TODO FIX ME!!!
-// TEST_CASE("GIVEN two units with different exponents AND different ratios WHEN
-// "
-//           "divided THEN result is unit with exponents subtracted AND ratio "
-//           "is gcd ") {
-//   constexpr SI::detail::unit_t<'X', 2, int64_t, std::ratio<1>> v1{2000};
-//   constexpr SI::detail::unit_t<'X', 1, int64_t, std::milli> v2{1000};
+TEST_CASE("GIVEN two units with different exponents AND different ratios WHEN  "
+          "divided THEN result is unit with exponents subtracted AND ratio "
+          "is divided ") {
+  constexpr SI::detail::unit_t<'X', 2, int64_t, std::ratio<1>> v1{4};
+  constexpr SI::detail::unit_t<'X', 1, int64_t, std::milli> v2{2};
 
-//   constexpr auto result = v1 / v2;
-//   STATIC_REQUIRE(std::is_same<
-//                  decltype(result),
-//                  const SI::detail::unit_t<'X', 1, int64_t,
-//                  std::milli>>::value);
-//   STATIC_REQUIRE(result.raw_value() == 2000);
-// }
+  constexpr auto result = v1 / v2;
+  STATIC_REQUIRE(std::is_same<
+                 decltype(result),
+                 const SI::detail::unit_t<'X', 1, int64_t, std::kilo>>::value);
+  STATIC_REQUIRE(result.raw_value() == 2);
+}
 
 TEST_CASE("GIVEN two units with different exponents AND same ratios WHEN "
           "divided THEN result is unit with exponents subtracted AND ratio "